Refactor and create file structure for pages
